refactor(frontend): consolidate orders workspace hook
