+++ frontend/src/pages/sales/__tests__/OrdersPage.filterbar.test.tsx
@@ -38,13 +38,46 @@ vi.mock('../components/OrderContextHeader', () => ({ default: () => <div>OrderCo
 vi.mock('../components/OrdersTable', () => ({ default: () => <div>OrdersTable</div> }))
 vi.mock('../components/OrderWorkspaceCard', () => ({ default: () => <div>OrderWorkspaceCard</div> }))
 vi.mock('../components/OrdersDialogs', () => ({ default: () => <div>OrdersDialogs</div> }))
-vi.mock('../hooks/ordersActions', () => ({
-  useOrdersActions: () => ({
-    handleEditOrder: vi.fn(),
+vi.mock('../hooks/useOrdersWorkspace', () => ({
+  useOrdersWorkspace: () => ({
+    searchInputRef: { current: null },
+    setShouldPreserveSearchFocus: vi.fn(),
+    focusedOrderIndex: -1,
+    orderListRef: { current: null },
+    viewDialog: false,
+    setViewDialog: vi.fn(),
+    blockedDialogOpen: false,
+    setBlockedDialogOpen: vi.fn(),
+    blockedDialogAction: 'edit',
+    deletedOrdersDialogOpen: false,
+    setDeletedOrdersDialogOpen: vi.fn(),
+    deleteConfirmOpen: false,
+    orderToDelete: null,
+    orderToDeleteName: '',
+    printDialogOpen: false,
+    setPrintDialogOpen: vi.fn(),
+    paymentDialogOpen: false,
+    setPaymentDialogOpen: vi.fn(),
+    isLoading: false,
+    journalEntryRef: null,
+    journalEntryRefLoading: false,
+    handleOrderSelect: vi.fn(),
+    handleNavigateUp: vi.fn(),
+    handleNavigateDown: vi.fn(),
+    handleNavigateToFirst: vi.fn(),
+    handleNavigateToLast: vi.fn(),
+    handlePageUpNavigation: vi.fn(),
+    handlePageDownNavigation: vi.fn(),
+    handleNavigateToInvoice: vi.fn(),
+    handleNavigateToPayment: vi.fn(),
+    navigateToJournalEntry: vi.fn(),
     handleOrderAction: vi.fn(),
-    handleRefundOrder: vi.fn(),
+    handleConfirmDelete: vi.fn(),
+    handleCancelDelete: vi.fn(),
+    handleEditOrder: vi.fn(),
+    handleRecordPayments: vi.fn(),
     handleUnpayOrder: vi.fn(),
-    openPaymentDialog: vi.fn(),
+    handleRefundOrder: vi.fn(),
     handleFulfillOrder: vi.fn(),
     handleUnfulfillOrder: vi.fn(),
     handleUnfulfillAndEdit: vi.fn(),
@@ -52,22 +85,7 @@ vi.mock('../hooks/ordersActions', () => ({
     handleUnpayAndEdit: vi.fn(),
     handleUnfulfillAndDelete: vi.fn(),
     handleUnpayAndDelete: vi.fn(),
-    handleConfirmDelete: vi.fn(),
-  }),
-}))
-vi.mock('../hooks/ordersSelection', () => ({
-  useOrdersSelection: () => ({
-    handleNavigateUp: vi.fn(),
-    handleNavigateDown: vi.fn(),
-    handleEnterAction: vi.fn(),
-    handlePageUpNavigation: vi.fn(),
-    handlePageDownNavigation: vi.fn(),
-    handleNavigateToFirst: vi.fn(),
-    handleNavigateToLast: vi.fn(),
-    handleEscapeAction: vi.fn(),
-    handleOrderSelect: vi.fn(),
-    handleNavigateToInvoice: vi.fn(),
-    handleNavigateToPayment: vi.fn(),
+    openPaymentDialog: vi.fn(),
   }),
 }))
 vi.mock('@/hooks/useNotification', () => ({
